docker/docker-tmpl/requirements.txt
docker
Jinja2
absl-py
